Fix typos, add required package, imports and methods, move enum Role into separate file, and keep class declaration format

+++ src/main/java/cn/edu/xjtlu/readingnotes/model/ReadingLog_.java
@@ -7,11 +7,11 @@ import jakarta.persistence.metamodel.StaticMetamodel;
 
 @StaticMetamodel(ReadingLog.class)
 public class ReadingLog_ {
-    public static volatile SingularAttribute<ReadingLog, Integer> id;
+    public static volatile SingularAttribute<ReadingLog, Long> id;
     public static volatile SingularAttribute<ReadingLog, String> title;
     public static volatile SingularAttribute<ReadingLog, String> author;
     public static volatile SingularAttribute<ReadingLog, LocalDate> date;
     public static volatile SingularAttribute<ReadingLog, Integer> spentTime;
     public static volatile SingularAttribute<ReadingLog, String> note;
-    public static volatile SingularAttribute<ReadingLog, Integer> userId;
+    public static volatile SingularAttribute<ReadingLog, Long> userId;
 }
